+++ graphql_client_codegen/src/interfaces.rs
@@ -12,14 +12,34 @@ use unions::union_variants;
 #[derive(Debug, Clone, PartialEq)]
 pub struct GqlInterface {
     pub description: Option<String>,
+    /// The set of object types implementing this interface.
     pub implemented_by: HashSet<String>,
+    /// The name of the interface. Should match 1-to-1 to its name in the GraphQL schema.
     pub name: String,
+    /// The interface's fields. Analogous to object fields.
     pub fields: Vec<GqlObjectField>,
     pub is_required: Cell<bool>,
 }
 
 impl GqlInterface {
-    pub fn new(name: Cow<str>, description: Option<&str>) -> GqlInterface {
+    /// filters the selection to keep only the fields that refer to the interface's own.
+    fn object_selection(&self, selection: &Selection) -> Selection {
+        Selection(
+            selection
+                .0
+                .iter()
+                // Only keep what we can handle
+                .filter(|f| match f {
+                    SelectionItem::Field(f) => f.name != "__typename",
+                    SelectionItem::FragmentSpread(_) => true,
+                    SelectionItem::InlineFragment(_) => false,
+                }).map(|a| (*a).clone())
+                .collect(),
+        )
+    }
+
+    /// Create an empty interface. This needs to be mutated before it is useful.
+    pub(crate) fn new(name: Cow<str>, description: Option<&str>) -> GqlInterface {
         GqlInterface {
             description: description.map(|d| d.to_owned()),
             name: name.into_owned(),
@@ -29,6 +49,38 @@ impl GqlInterface {
         }
     }
 
+    /// The generated code for each of the selected field's types. See [shared::field_impls_for_selection].
+    pub(crate) fn field_impls_for_selection(
+        &self,
+        context: &QueryContext,
+        selection: &Selection,
+        prefix: &str,
+    ) -> Result<Vec<TokenStream>, failure::Error> {
+        ::shared::field_impls_for_selection(
+            &self.fields,
+            context,
+            &self.object_selection(selection),
+            prefix,
+        )
+    }
+
+    /// The code for the interface's corresponding struct's fields.
+    pub(crate) fn response_fields_for_selection(
+        &self,
+        context: &QueryContext,
+        selection: &Selection,
+        prefix: &str,
+    ) -> Result<Vec<TokenStream>, failure::Error> {
+        response_fields_for_selection(
+            &self.name,
+            &self.fields,
+            context,
+            &self.object_selection(selection),
+            prefix,
+        )
+    }
+
+    /// Generate all the code for the interface.
     pub(crate) fn response_for_selection(
         &self,
         query_context: &QueryContext,
@@ -42,19 +94,6 @@ impl GqlInterface {
             .extract_typename()
             .ok_or_else(|| format_err!("Missing __typename in selection for {}", prefix))?;
 
-        let object_selection = Selection(
-            selection
-                .0
-                .iter()
-                // Only keep what we can handle
-                .filter(|f| match f {
-                    SelectionItem::Field(f) => f.name != "__typename",
-                    SelectionItem::FragmentSpread(_) => true,
-                    SelectionItem::InlineFragment(_) => false,
-                }).map(|a| (*a).clone())
-                .collect(),
-        );
-
         let union_selection = Selection(
             selection
                 .0
@@ -67,16 +106,10 @@ impl GqlInterface {
                 .collect(),
         );
 
-        let object_fields = response_fields_for_selection(
-            &self.name,
-            &self.fields,
-            query_context,
-            &object_selection,
-            prefix,
-        )?;
+        let object_fields =
+            self.response_fields_for_selection(query_context, &selection, prefix)?;
 
-        let object_children =
-            field_impls_for_selection(&self.fields, query_context, &object_selection, prefix)?;
+        let object_children = self.field_impls_for_selection(query_context, &selection, prefix)?;
         let (mut union_variants, union_children, used_variants) =
             union_variants(&union_selection, query_context, prefix)?;